requirements.txt
fastapi
uvicorn
websockets
openai
python-multipart
aiofiles
python-dotenv
pydantic-settings